Removed folder lib, moved modules/packges to core
Previously had used pairs/lib as a package repository, decided it would
be better to put the code in pairs/core, this seems to be the intent
of the cement framework.

The only import statement that had to be changed was in the base
controller, all of the imports in the former lib folder were relative.