Added configure command, setup config file
A config file can now be generated with the command `configure`. The
default config file location is set as '~/.config/pairs.ini'. If this
file does not exist, app default settings are used (these defaults are
set up in base.py). If this file does exist, the defaults are overidden
(i.e. the user configs are used).