Changed config file format to yml ins. ini
Tried to use ini file as format for config, but cement base is set to
yml, so this makes more sense. Before this commit an empty file None was
being written in the project folder for some reason. Instead of fixing
that, just changed to the defaults.

Also changed the default config path from '~/.conf/pairs.yml' to
'~/.conf/pairs/pairs.yml', which is also compatible with the cement
default.